modifications for practicing with metronome
double click allows metronome letters to be deleted. however, practice letters were not updated to show this.

made it so that practice letters do not update if metronome is displayed. matches use metronome letters instead of practice letters, so that match can be determined whether or not metronome is displayed.

reset lesson whenever metronome display button is pressed